+++ app/src/main/java/com/jobinlawrance/weather/data/network/model/Main.kt
@@ -3,7 +3,7 @@ package com.jobinlawrance.weather.data.network.model
 data class Main(
 	val temp: Double? = null,
 	val tempMin: Double? = null,
-	val humidity: Int? = null,
-	val pressure: Int? = null,
+	val humidity: Double? = null,
+	val pressure: Double? = null,
 	val tempMax: Double? = null
 )
