+++ app/src/main/java/com/jobinlawrance/weather/data/network/model/Sys.kt
@@ -2,8 +2,8 @@ package com.jobinlawrance.weather.data.network.model
 
 data class Sys(
 	val country: String? = null,
-	val sunrise: Int? = null,
-	val sunset: Int? = null,
+	val sunrise: Long? = null,
+	val sunset: Long? = null,
 	val id: Int? = null,
 	val type: Int? = null,
 	val message: Double? = null
